[1/2] Allow homeservers to send registration emails | Sending the email (#5835)
~~Fixes https://github.com/matrix-org/synapse/issues/5833~~ Moved out to ~~https://github.com/matrix-org/synapse/pull/5863~~
Part of fixing https://github.com/matrix-org/synapse/issues/5751

Decouples the activity of sending registration emails and binding them to an identity server.

This PR simply sends the registration email, but clicking it does not approve the user for registration. That will come in PR #2.

Some of this makes use of existing stuff for sending password reset emails from Synapse. Some work was done to make that stuff even more generic.

Upgrade notes:
* There is a new top-level config option, `account_threepid_delegate` which defines the address of an identity server that you would like to send registration/password reset emails on your behalf.

  The option `email.trust_identity_server_for_password_resets` has been replaced by this. If you set `email.trust_identity_server_for_password_resets` in your config to `true`, please remove it and configure `account_threepid_delegate` instead. The [sample config](https://github.com/matrix-org/synapse/blob/anoa/reg_email_sending_email/docs/sample_config.yaml) has information on how to configure it.

Note: This PR does not allow homeservers to send emails when simply adding an email to your account. That will come after this and will be blocked on a new MSC.

Part [2/2] will be successfully completing the registration step when `/submit_token` is hit with the correct details, and clearing out the `password_servlet` flag stuff, which is no longer needed. Will be a much smaller PR than this one.

~~Requires https://github.com/matrix-org/synapse/pull/5863~~ has been merged into the base branch.
~~Requires https://github.com/matrix-org/synapse/pull/5876~~ has been merged into the base branch.