Fix demo script on ipv6-supported boxes
The synapse demo was a bit flakey in terms of supporting federation. It turns
out that if your computer resolved localhost to ::1 instead of 127.0.0.1, the
built-in federation blacklist specified in start.sh would still block it, since
it contained an entry for ::/127. Removing this no longer prevents Synapse from
contacting ::1, federation works again on these boxes.