login: Fix bug in automatic JWT user creation
When a previously non-existent user is authenticated with m.login.jwt
(after the token has been validated), the user should be created. But a
bug made checking if the user exists reset the claimed username to None,
and we would see generic TypeError exceptions when we asked synapse to
process it further:

  ...
  File "/usr/lib/python3/dist-packages/synapse/rest/client/v1/login.py", line 393, in do_jwt_login
    result = await self._complete_login(
  File "/usr/lib/python3/dist-packages/synapse/rest/client/v1/login.py", line 338, in _complete_login
    localpart=UserID.from_string(user_id).localpart
  File "/usr/lib/python3/dist-packages/synapse/types.py", line 171, in from_string
    if len(s) < 1 or s[0:1] != cls.SIGIL:
  TypeError: object of type 'NoneType' has no len()

This regression was introduced in 541f1b92d (part of the v1.6.0 release
of synapse); it only affects auth flows where _complete_login is called
with create_non_existant_user=True, which turns out is only the
m.login.jwt auth flow.

Signed-off-by: Olof Johansson <olof@ethup.se>