Database migration to add 'service' to 'threepid_validation_session'
We have a problem whereby during threepid (email) validation we need to know whether to handle the validation token ourselves or via an account threepid delegate.

Earlier we could just check whether we were delegating email or not, but now that email can be either Synapse or a threepid delegate depending on which service (adding threepids, password reset) the validation is for, we need a way to find that out.

So we store that information along with the validation session in the DB.