Check the validation session's service type when a token is submitted.
If a service is specified, check with our new `delegate_for` option for
whether this should be handled by an account threepid delegate, or
remotely.

If there isn't a threepid session found with the details given by the
client, assume that this is a remote session.

Also modify the `is_enabled` method to check both `add_threepid` and `password_reset` services. If any of them are configured, then this checker is enabled.

And again, refactor a little bit by pulling config option polling into the constructor.