Break up threepid_behaviour_email into add_threepid/password_reset
So we initially had a variable called `threepid_behaviour_email`
attached to the config object which was an instance of
ThreepidBehaviour, the latter having three possible values - REMOTE,
LOCAL or OFF. The rest of the codebase would basically use this to
determine whether to try and do email stuff itself, or to delegate the
operation to a third-party service (an identity server usually).

Well, now that the sysadmin can choose whether to do this for adding a
threepid vs. password reset, we now need two ThreepidBehaviour values.
The code can then check either the one for adding a threepid, or the one
for password reset, depending on what it's doing.

This commit is breaking that up and also cleaning up the email template
loading code such that templates are only loaded if necessary i.e. we'll
be sending emails with that template.

Oh, we also introduce ThreepidService, which is another enum like
ThreepidBehaviour, but this one specifies which service you're
validating (whether it's password reset or adding a threepid). I chose
and enum again instead of a straight string as then we can easily avoid
typos.