Generate invite rejections locally on failure.
This is the core change which fixes #2181. The basic premise is that, when we
fail to reject an invite via the remote server, we can generate our own
out-of-band leave event and persist it as an outlier, so that we have something
to send to the client.

This makes the old `locally_reject_invite` impl redundant; a future commit will
remove it.