Process the cache stream first for incoming replication
Receiving new data about e.g. events, read receipts, etc., is likely to
result in some database actions. An example of that is that we send a
push on new read receipts, which involves calculating a badge from
unread counts. Because of this, we probably want to process cache
updates/invalidations first so these actions are not using outdated
cache data. In the previous example, we want to make sure we invalidate
the cache on calculating the unread count for the user/room the read
receipt is for before sending a push.