Move building typing edu to federation senders.
Instead of always building EDUs on the writer process and using the
federation send queue to stream to the federation senders, we manage
sending of EDUs on the federation senders by looking at typing
replication stream.

This means we also have to move the timers for federation pokes to the
federation sending processes too.