Fix bug which could cause replication command processing to wedge
`_process_command` was not correctly wraped in try/except, so if it threw an
exception, we wouldn't go on to process the rest of the queued commands, which
would mean that every subsequent command would cause an assertion failure.

This reshuffles the logic so that there is only one call to `_process_command`.